refactor: adjusting several names
